Restructure peer_review JS into core/roles/ui subfolders
Group the flat module pile by concern for navigability:
  core/   peer_review.js, state_current_review.js, selectors.js, store.js
  roles/  opr_reviewer.js, opr_reviewer_logic.js, opr_contributor.js
  ui/     navigation.js, summary.js, field_description.js, field_history.js, input_toggles.js
  data/   schema.json, example_datenmodell.json, template_datenmodell.json
  __tests__/  *.test.js
main.js, api.js, utilities.js stay at the root. All relative imports were
rewritten accordingly; main.js (the vite entry) keeps its path so the
vite_asset reference is unchanged. Also updates the REUSE.toml annotation
paths for the moved JSON files and the profile dashboard's opr_reviewer.js
script src. No code behaviour change.
