+++ dataedit/static/peer_review/core/selectors.js
