+++ dataedit/static/peer_review/main.js
@@ -4,14 +4,14 @@
 // SPDX-FileCopyrightText: 2026 Jonas Huber <https://github.com/jh-RLI> © Reiner Lemoine Institut
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import * as common from "./peer_review.js";
-import { selectState } from './peer_review.js';
-import { selectNextField, selectPreviousField, selectFirstReviewableField, selectFirstContributorField } from './navigation.js';
-import { setGetFieldState } from './state_current_review.js';
+import * as common from "./core/peer_review.js";
+import { selectState } from './core/peer_review.js';
+import { selectNextField, selectPreviousField, selectFirstReviewableField, selectFirstContributorField } from './ui/navigation.js';
+import { setGetFieldState } from './core/state_current_review.js';
 
 // Static imports avoid the "Failed to fetch" dynamic import errors
-import { initReviewer } from './opr_reviewer.js';
-import { initContributor } from './opr_contributor.js';
+import { initReviewer } from './roles/opr_reviewer.js';
+import { initContributor } from './roles/opr_contributor.js';
 
 // Expose functions to global window scope for HTML onclick events
 window.selectState = selectState;
