+++ dataedit/static/peer_review/ui/field_description.js
@@ -10,7 +10,7 @@
 // Extracted from peer_review.js as part of the Phase 3 frontend split;
 // peer_review.js re-exports these so existing importers keep working.
 
-import { escapeHtml } from "./utilities.js";
+import { escapeHtml } from "../utilities.js";
 
 // --- DOM helpers ---
 function getFieldElByKey(fieldKey) {
