+++ dataedit/static/peer_review/ui/input_toggles.js
