+++ dataedit/static/peer_review/ui/summary.js
@@ -4,14 +4,14 @@
 // SPDX-FileCopyrightText: 2026 Vismaya Jochem <https://github.com/vismayajochem> © Reiner Lemoine Institut
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import { current_review, selectedState } from "./peer_review.js";
-import { getFieldState } from "./state_current_review.js";
+import { current_review, selectedState } from "../core/peer_review.js";
+import { getFieldState } from "../core/state_current_review.js";
 import {
   isEmptyValue,
   isEffectivelyEmpty,
   sendJson,
   escapeHtml,
-} from "./utilities.js";
+} from "../utilities.js";
 import { updatePercentageDisplay } from "./navigation.js";
 export function renderSummaryPageFields() {
   const acceptedFields = [];
