+++ login/templates/login/user_review.html
@@ -515,7 +515,7 @@
       review_id: "{{ latest_review_id }}",
     };
   </script>
-  <script src="{% static '/peer_review/opr_reviewer.js' %}"></script>
+  <script src="{% static '/peer_review/roles/opr_reviewer.js' %}"></script>
   <script>
     $(function () {
       $('[data-bs-toggle="popover"]').popover();
